+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- added: (zkSync) Add USDC
+- fixed: (zkSync) Fix USDC.e currency code
+
 ## 4.17.1 (2024-07-30)
 
 - fixed: (ETH) Fix incorrect protocol in blockbook-ws connection URLs